compositor-drm: Set higher precision gamma LUT
Use the fancier new gamma LUT thing and set it up apporpriately.
Ideally in the case where we apply gamma with shaders we would
entirely disable the crtc gamma. But the uapi doens't provide
a way to do that :( So at least we can avoid the quality loss
from the legacy gamma by programming a linear gamma into the
higher precision LUT. Even there we lose precision needlessly
to the degmama LUT on platforms with a split gamma LUT. The
kernel uapi for this stuff definitely needs more thought...

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>