gl-renderer: Perform colorspace conversion
The client's colorspace may not match the colorspace we want to
output to the display. Perform the conversion with shaders.

The conversion follows this sequence:
 RGB -> XYZ -> LMS -> CAT -> LMS -> XYZ -> RGB

Where CAT is the von Kries chromatic adaptation to deal with
potentially different white points of the two colorspaces.
We're using the Bradford matrix for XYZ<->LMS.

Fortunately all of those are just 3x3 matrices so
it all collapses down into a single matrix.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>